ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.10